Rename walk to walk_functions.py
