+++ synapseutils/walk_functions.py
